Implement RISC-V QoS extension (Ssqosid)
Implement the sqoscfg CSR and tie it to the Ssqosid extension. The CSR
is used to set IDs for resource control (RCID) and monitoring counters
(MCID).

The CSR is defined in the S-mode space, but the spec states that
accessing it when V=1 shall cause a virtual instruction exception.
Implement this by calling the hmode predicate.

Signed-off-by: Kornel Dulęba <mindal@semihalf.com>
[dfustini: rebase on v8.0.0-rc1, reword commit message]
Signed-off-by: Drew Fustini <dfustini@baylibre.com>